Extract common view methods into BaseViewTrait
